+++ R/mod_addComment.R
@@ -32,7 +32,7 @@ addCommentUI <- function(id) {
 #' 
 addCommentServer <- function(id, metric_abrv, user, credentials, pkg_name) {
   if (missing(credentials))
-    credentials <- get_golem_config("credentials", file = app_sys("db-config.yml"))
+    credentials <- get_db_config("credentials")
   
   moduleServer(id, function(input, output, session) {
     
@@ -41,7 +41,8 @@ addCommentServer <- function(id, metric_abrv, user, credentials, pkg_name) {
       metric_name <- switch(metric_abrv,
                             mm = "Maintenance Metrics",
                             cum = "Community Usage Metrics",
-                            se = "Source Explorer")
+                            se = "Source Explorer",
+                            fe = "Function Explorer")
       
       textAreaInput(
         session$ns("add_comment"),