+++ R/mod_communityMetrics.R
@@ -25,7 +25,7 @@ communityMetricsUI <- function(id) {
 #' 
 communityMetricsServer <- function(id, selected_pkg, community_metrics, user, credentials) {
   if (missing(credentials))
-    credentials <- get_golem_config("credentials", file = app_sys("db-config.yml"))
+    credentials <- get_db_config("credentials")
   
   moduleServer(id, function(input, output, session) {