+++ R/mod_metricGrid.R
@@ -21,7 +21,7 @@ metricGridUI <- function(id) {
 metricGridServer <- function(id, metrics) {
   moduleServer(id, function(input, output, session) {
     
-    metric <- dbSelect("select * from metric", db_name = golem::get_golem_options('assessment_db_name'))
+    metric_tbl <- dbSelect("select * from metric", db_name = golem::get_golem_options('assessment_db_name'))
     
     output$grid <- renderUI({
       req(nrow(metrics()) > 1) # need at least two cards to make a metric grid UI
@@ -32,25 +32,27 @@ metricGridServer <- function(id, metrics) {
       fluidRow(style = "padding-right: 10px", class = "card-group",
                map(column_vector_grid_split, 
                    ~ column(width= 4,map(.x,~ metricBoxUI(session$ns(metrics()$name[.x]))))),
-      if(any(!(metrics()$title %in% metric$long_name)) & stringr::str_extract(session$ns(id), "\\w+") != "databaseView") {
+      if(any(!(metrics()$title %in% metric_tbl$long_name)) & stringr::str_extract(session$ns(id), "\\w+") != "databaseView") {
         tags$em("* Provided for additional context. Not a {riskmetric} assessment, so this measure will not impact the risk score.")
       } 
       )
     })
     
+    
     observeEvent(req(nrow(metrics()) > 0), {
-      apply(metrics(), 1, function(metric)
-        metricBoxServer(id = metric['name'],
-            title = metric['title'],
-            desc = metric['desc'],
-            value = dplyr::case_when(metric['name'] != 'has_bug_reports_url' ~ metric['value'],
-                                     metric['value'] == "1" ~ 'TRUE',
+      apply(metrics(), 1, function(m)
+        metricBoxServer(id = m['name'],
+            title = m['title'],
+            desc = m['desc'],
+            value = dplyr::case_when(m['name'] != 'has_bug_reports_url' ~ m['value'],
+                                     m['value'] == "1" ~ 'TRUE',
                                      TRUE ~ 'FALSE'),
-            is_perc = metric['is_perc'] == 1,
-            is_url = metric['is_url'] == 1,
-            succ_icon = metric['succ_icon'],
-            icon_class = metric['icon_class'],
-            type = metric['type']
+            score = m['score'],
+            is_perc = m['is_perc'] == 1,
+            is_url = m['is_url'] == 1,
+            succ_icon = m['succ_icon'],
+            icon_class = m['icon_class'],
+            type = m['type']
           )
         )
     })