+++ R/mod_reweightView.R
@@ -15,20 +15,18 @@ reweightViewUI <- function(id) {
 #' @param id the module id
 #' @param user the user name
 #' @param decision_list the list containing the decision automation criteria
-#' @param trigger_events a reactive values object to trigger actions here or elsewhere
 #' 
 #' 
 #' @import dplyr
 #' @importFrom DT datatable formatStyle styleEqual renderDataTable
 #' @importFrom shinyjs enable disable delay
-#' @importFrom shinydashboard box
 #' @importFrom DBI dbConnect dbDisconnect
 #' @importFrom RSQLite SQLite sqliteCopyDatabase
 #' 
 #' @keywords internal
-reweightViewServer <- function(id, user, decision_list, credentials, trigger_events) {
+reweightViewServer <- function(id, user, decision_list, credentials) {
   if (missing(credentials))
-    credentials <- get_golem_config("credentials", file = app_sys("db-config.yml"))
+    credentials <- get_db_config("credentials")
   moduleServer(id, function(input, output, session) {
     
     exportTestValues(
@@ -88,59 +86,63 @@ reweightViewServer <- function(id, user, decision_list, credentials, trigger_eve
       tagList(
         tags$section(
           br(), br(),
-          shinydashboard::box(width = 12, status = "primary",
-              title = h2("View/Change Weights", style = "margin-top: 5px", align = "center"),
-              solidHeader = TRUE,
-              br(),
-              fluidRow(
-                column(width = 5, offset = 5, align = "left",
-                       h3("Set new weights:"),
-                )),
-              fluidRow(
-                column(width = 2, offset = 5, align = "left",
-                       selectInput(NS(id, "metric_name"), "Select metric", curr_new_wts()$name, selected = curr_new_wts()$name[1]) ),
-                column(width = 2, align = "left",
-                       numericInput(NS(id, "metric_weight"), "Choose new weight", min = 0, value = curr_new_wts()$new_weight[1]) ),
-                column(width = 1,
-                       br(),
-                       actionButton(NS(id, "update_weight"), "Confirm", class = "btn-secondary") ) ),
-              br(), br(), 
-              fluidRow(
-                column(width = 3, offset = 1, align = "center",
-                       
-                       br(), br(), br(), 
-                       tags$hr(class = "hr_sep"),
-                       br(), br(),
-                       
-                       h3("Download database"),
-                       downloadButton(NS(id, "download_database_btn"),
-                                      "Download",
-                                      class = "btn-secondary"),
-                       
-                       br(), br(), br(), 
-                       tags$hr(class = "hr_sep"),
-                       br(), br(),
-                       
-                       h3("Apply new weights and re-calculate risk for each package"),
-                       actionButton(NS(id, "update_pkg_risk"), "Update", class = "btn-secondary")
-                       
-                ),
-                column(width = 6, style = "border: 1px solid rgb(77, 141, 201)",
-                       offset = 1,
-                       h3("Current Risk Score Weights by Metric", align = "center"),
-                       DT::dataTableOutput(NS(id, "weights_table")))
+          div(class = c("box", "box-primary", "box-solid"),
+              h3(class = "box-title",
+                 h2("View/Change Weights", style = "margin-top: 5px", align = "center")
               ),
-              br(), br(), br(),
-              fluidRow(
-                column(width = 1),
-                column(width = 10,
-                       h5(em("Note: Changing the weights of the metrics will not update the
+              div(class = "box-body",
+                  br(),
+                  fluidRow(
+                    column(width = 5, offset = 5, align = "left",
+                           h3("Set new weights:"),
+                    )),
+                  fluidRow(
+                    column(width = 2, offset = 5, align = "left",
+                           selectInput(NS(id, "metric_name"), "Select metric", curr_new_wts()$name, selected = curr_new_wts()$name[1]) ),
+                    column(width = 2, align = "left",
+                           numericInput(NS(id, "metric_weight"), "Choose new weight", min = 0, value = curr_new_wts()$new_weight[1]) ),
+                    column(width = 1,
+                           br(),
+                           actionButton(NS(id, "update_weight"), "Confirm", class = "btn-secondary") ) ),
+                  br(), br(), 
+                  fluidRow(
+                    column(width = 3, offset = 1, align = "center",
+                           
+                           br(), br(), br(), 
+                           tags$hr(class = "hr_sep"),
+                           br(), br(),
+                           
+                           h3("Download database"),
+                           downloadButton(NS(id, "download_database_btn"),
+                                          "Download",
+                                          class = "btn-secondary"),
+                           
+                           br(), br(), br(), 
+                           tags$hr(class = "hr_sep"),
+                           br(), br(),
+                           
+                           h3("Apply new weights and re-calculate risk for each package"),
+                           actionButton(NS(id, "update_pkg_risk"), "Update", class = "btn-secondary")
+                           
+                    ),
+                    column(width = 6, style = "border: 1px solid rgb(77, 141, 201)",
+                           offset = 1,
+                           h3("Current Risk Score Weights by Metric", align = "center"),
+                           DT::dataTableOutput(NS(id, "weights_table")))
+                  ),
+                  br(), br(), br(),
+                  fluidRow(
+                    column(width = 1),
+                    column(width = 10,
+                           h5(em("Note: Changing the weights of the metrics will not update the
                risk of the packages on the database until 'Update' button is selected.
                ")), align = "center"),
-                column(width = 1)
-              ),
-              br(), br()
-          )
+                    column(width = 1)
+                  ),
+                  br(), br()
+              )
+          ) %>%
+            column(width = 12)
         )
       )
     })
@@ -244,7 +246,7 @@ reweightViewServer <- function(id, user, decision_list, credentials, trigger_eve
       req("weight_adjust" %in% credentials$privileges[[user$role]])
       removeModal()
       
-      trigger_events[["reset_pkg_upload"]] <- trigger_events[["reset_pkg_upload"]] + 1
+      session$userData$trigger_events[["reset_pkg_upload"]] <- session$userData$trigger_events[["reset_pkg_upload"]] + 1
       
       # Update the weights in the `metric` table to reflect recent changes
       # First, which weights are different than the originals?
@@ -263,7 +265,7 @@ reweightViewServer <- function(id, user, decision_list, credentials, trigger_eve
           dplyr::mutate(new_weight = weight)
       )
       
-      trigger_events$reset_sidebar <- trigger_events$reset_sidebar + 1
+      session$userData$trigger_events$reset_sidebar <- session$userData$trigger_events$reset_sidebar + 1
       
       # update for each package
       all_pkgs <- dbSelect("SELECT DISTINCT name AS pkg_name FROM package")
@@ -328,7 +330,7 @@ reweightViewServer <- function(id, user, decision_list, credentials, trigger_eve
     output$download_database_btn <- downloadHandler(
       
       filename = function() {
-        glue::glue("datase_backup-{Sys.Date()}.sqlite")
+        glue::glue("datase_backup-{get_Date()}.sqlite")
       },
       content = function(file) {
         con <- DBI::dbConnect(RSQLite::SQLite(), golem::get_golem_options('assessment_db_name'))